iOS tests were redesigned
