+++ src/test/java/io/appium/java_client/ios/IOSElementTest.java
@@ -21,7 +21,7 @@ import static org.junit.Assert.assertNotEquals;
 
 import org.junit.Test;
 
-public class IOSElementTest extends BaseIOSTest {
+public class IOSElementTest extends AppIOSTest {
 
     @Test public void findByAccessibilityIdTest() {
         assertNotEquals(driver.findElementsByClassName("UIAWindow").get(1)
