+++ src/test/java/io/appium/java_client/ios/IOSSwipeGestureTest.java
@@ -7,59 +7,13 @@ import static org.junit.Assert.assertTrue;
 
 import io.appium.java_client.MobileElement;
 import io.appium.java_client.SwipeElementDirection;
-import io.appium.java_client.remote.IOSMobileCapabilityType;
-import io.appium.java_client.remote.MobileCapabilityType;
 import io.appium.java_client.service.local.AppiumDriverLocalService;
-import io.appium.java_client.service.local.AppiumServerHasNotBeenStartedLocallyException;
 import org.junit.After;
-import org.junit.AfterClass;
-import org.junit.BeforeClass;
 import org.junit.Test;
 import org.openqa.selenium.Point;
 import org.openqa.selenium.interactions.internal.Coordinates;
-import org.openqa.selenium.remote.DesiredCapabilities;
 
-import java.io.File;
-
-public class IOSSwipeGestureTest {
-
-    private static AppiumDriverLocalService service;
-    protected static IOSDriver<IOSElement> driver;
-
-    @BeforeClass
-    public static void beforeClass() throws Exception {
-        service = AppiumDriverLocalService.buildDefaultService();
-        service.start();
-
-        if (service == null || !service.isRunning()) {
-            throw new AppiumServerHasNotBeenStartedLocallyException(
-                "An appium server node is not started!");
-        }
-
-        File appDir = new File("src/test/java/io/appium/java_client");
-        File app = new File(appDir, "UICatalog.app.zip");
-        DesiredCapabilities capabilities = new DesiredCapabilities();
-        capabilities.setCapability(MobileCapabilityType.BROWSER_NAME, "");
-        capabilities.setCapability(MobileCapabilityType.PLATFORM_VERSION, "9.2");
-        capabilities.setCapability(MobileCapabilityType.DEVICE_NAME, "iPhone Simulator");
-        //sometimes environment has performance problems
-        capabilities.setCapability(IOSMobileCapabilityType.LAUNCH_TIMEOUT, 500000);
-        capabilities.setCapability(MobileCapabilityType.APP, app.getAbsolutePath());
-        driver = new IOSDriver<>(service.getUrl(), capabilities);
-    }
-
-    /**
-     * finishing.
-     */
-    @AfterClass
-    public static void afterClass() {
-        if (driver != null) {
-            driver.quit();
-        }
-        if (service != null) {
-            service.stop();
-        }
-    }
+public class IOSSwipeGestureTest extends UICatalogIOSTest {
 
     @After public void afterMethod() {
         driver.resetApp();
